refactor db driver and calculator code.
