+++ calculator/constants.py
@@ -1,0 +1,2 @@
+DB_API_URL = "http://db-api:5000"
+WEB_URL = "http://web:8000"
