+++ db-api/Dockerfile
@@ -7,8 +7,7 @@ COPY db-api/requirements.txt ./requirements.txt
 
 RUN pip install --no-cache-dir -r requirements.txt
 
-COPY db-api/db.py ./db.py
-COPY db-api/main.py ./main.py
+COPY db-api/*.py ./
 COPY db-api/forge_data.json ./forge_data.json
 
 CMD ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "5000"]
