+++ web/Dockerfile
@@ -17,6 +17,6 @@ COPY web/requirements.txt ./requirements.txt
 RUN pip install --no-cache-dir -r requirements.txt
 
 COPY --from=frontend-build /frontend/dist ./static
-COPY web/main.py ./main.py
+COPY web/*.py ./
 
 CMD ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "8000"]
